excluded a couple tests; fixed alerts tests
